Refactor and cleanup solution
