+++ TagsCloud/TagsCloudVisualization.CLI/OptionsExtensions.cs
@@ -7,6 +7,7 @@ using TagsCloudDrawer.ImageSavior;
 using TagsCloudDrawer.ImageSettings;
 using TagsCloudVisualization.CloudLayouter;
 using TagsCloudVisualization.DrawerSettingsProvider;
+using TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator;
 
 namespace TagsCloudVisualization.CLI
 {
