+++ TagsCloud/TagsCloudVisualization/DrawerSettingsProvider/ITagDrawableSettingsProvider.cs
@@ -1,4 +1,6 @@
-﻿namespace TagsCloudVisualization.DrawerSettingsProvider
+﻿using TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator;
+
+namespace TagsCloudVisualization.DrawerSettingsProvider
 {
     public interface ITagDrawableSettingsProvider
     {
