+++ TagsCloud/TagsCloudVisualization/DrawerSettingsProvider/TagDrawableSettingsProvider.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using TagsCloudDrawer.ColorGenerators;
+using TagsCloudVisualization.DrawerSettingsProvider.TagColorGenerator;
 
 namespace TagsCloudVisualization.DrawerSettingsProvider
 {
